+++ src/components/post/Comments.js
@@ -3,15 +3,9 @@ import styled from "styled-components";
 
 import Comment from "./Comment";
 
-const CommentsSection = styled.section`
-  
-  margin-top: 10rem;
-  border-top: 1px solid ${(props) => props.theme.lightColor};
-`;
-
 export default function Comments({ comments, myInfo, setPostPageData }) {
   return (
-    <CommentsSection>
+    <section>
       <h4 className="sr-only">댓글 목록</h4>
       <ol>
         {comments.map((comment, index) => (
@@ -23,6 +17,6 @@ export default function Comments({ comments, myInfo, setPostPageData }) {
           />
         ))}
       </ol>
-    </CommentsSection>
+    </section>
   );
 }
