+++ src/components/post/Dialog.js
@@ -1,29 +1,11 @@
-import { useEffect, useState } from "react";
 import styled from "styled-components";
-import API from "../../utils/api";
 import Comments from "./Comments";
-import Writing from "./Writing";
 
 const DialogSection = styled.section`
   height: calc(100% - 406.58px);
 `;
 
-export default function Dialog({ comments, setPostPageData }) {
-  const [myInfo, setMyInfo] = useState("");
-  const getMyInfo = async () => {
-    const response = await API.get("/user/myinfo", {
-      Authorization: `Bearer ${localStorage.getItem("token")}`,
-    });
-
-    const responseData = await response.data;
-
-    setMyInfo(responseData);
-  };
-
-  useEffect(() => {
-    getMyInfo();
-  }, []);
-
+export default function Dialog({ comments, setPostPageData, myInfo }) {
   return (
     <>
       {comments && myInfo ? (
@@ -33,11 +15,6 @@ export default function Dialog({ comments, setPostPageData }) {
             comments={comments}
             myInfo={myInfo}
             setPostPageData={setPostPageData}
-          />
-          <Writing
-            comments={comments}
-            setPostPageData={setPostPageData}
-            myInfo={myInfo}
           />
         </DialogSection>
       ) : (
